Removed redundant casts for methods which were switched from using MathUtil.clamp(double,double,double) to MathUtil.clamp(int,int,int)

+++ simulations-java/common/controls/src/edu/colorado/phet/common/controls/SpectrumSliderWithSquareCursor.java
@@ -187,7 +187,7 @@ public class SpectrumSliderWithSquareCursor extends CompositePhetGraphic {
     public void setValue( int value ) {
 
         // Silently clamp the value to the allowed range.
-        _value = (int) MathUtil.clamp( _minimum, value, _maximum );
+        _value = MathUtil.clamp( _minimum, value, _maximum );
 
         // Update the knob.
         updateKnob();
